Improvements for links and centralized code
