+++ src/actions/AssetActionPlugin.js
@@ -3,9 +3,8 @@ import ActionPlugin from './ActionPlugin';
 export default class AssetActionPlugin extends ActionPlugin {
 
   constructor(asset, component, id) {
-    super(id);
+    super(id, component);
     this.asset = asset;
-    this.component = component;
   }
 
 }
