+++ src/components/Asset.vue
@@ -20,25 +20,7 @@
     <b-collapse :id="uid" v-model="expanded" :accordion="type" role="tabpanel" @input="collapseToggled">
       <b-card-body>
         <b-card-title><span v-html="fileFormat" /></b-card-title>
-        <b-button-group class="actions" v-if="href">
-          <b-button v-if="isBrowserProtocol" :href="href" target="_blank" variant="primary">
-            <b-icon-box-arrow-up-right v-if="browserCanOpenFile" /> 
-            <b-icon-download v-else />
-            {{ buttonText }}
-          </b-button>
-          <CopyButton variant="primary" :copyText="href">
-            {{ copyButtonText }}
-          </CopyButton>
-          <b-button v-if="canShow && !shown" @click="show" variant="primary">
-            <b-icon-eye class="mr-1" />
-            <template v-if="isThumbnail">{{ $t('assets.showThumbnail') }}</template>
-            <template v-else>{{ $t('assets.showOnMap') }}</template>
-          </b-button>
-          <b-button v-for="action of actions" v-bind="action.btnOptions" :key="action.id" variant="primary" @click="action.onClick">
-            <component v-if="action.icon" :is="action.icon" class="mr-1" />
-            {{ action.text }}
-          </b-button>
-        </b-button-group>
+        <HrefActions isAsset :data="asset" :shown="shown" @show="show" />
         <b-card-text class="mt-4" v-if="asset.description">
           <Description :description="asset.description" compact />
         </b-card-text>
@@ -49,27 +31,22 @@
 </template>
 
 <script>
-import { BCollapse, BIconBoxArrowUpRight, BIconCheck, BIconChevronRight, BIconChevronDown, BIconDownload, BIconEye } from 'bootstrap-vue';
+import { BCollapse, BIconCheck, BIconChevronRight, BIconChevronDown } from 'bootstrap-vue';
 import { formatMediaType } from '@radiantearth/stac-fields/formatters';
-import { mapGetters, mapState } from 'vuex';
+import { mapState } from 'vuex';
 import Description from './Description.vue';
-import STAC from '../models/stac';
-import Utils, { browserProtocols, imageMediaTypes, mapMediaTypes } from '../utils';
+import HrefActions from './HrefActions.vue';
 import StacFieldsMixin from './StacFieldsMixin';
-import AssetActions from '../../assetActions.config';
 
 export default {
   name: 'Asset',
   components: {
     BCollapse,
-    BIconBoxArrowUpRight,
     BIconCheck,
     BIconChevronDown,
     BIconChevronRight,
-    BIconDownload,
-    BIconEye,
-    CopyButton: () => import('./CopyButton.vue'),
     Description,
+    HrefActions,
     Metadata: () => import('./Metadata.vue')
   },
   mixins: [
@@ -123,12 +100,6 @@ export default {
   },
   computed: {
     ...mapState(['buildTileUrlTemplate', 'useTileLayerAsFallback', 'url', 'stateQueryParameters']),
-    ...mapGetters(['getRequestUrl']),
-    actions() {
-      return Object.entries(AssetActions)
-        .map(([id, plugin]) => new plugin(this.asset, this, id))
-        .filter(plugin => plugin.show);
-    },
     tileRendererType() {
       if (this.buildTileUrlTemplate && !this.useTileLayerAsFallback) {
         return 'server';
@@ -143,32 +114,6 @@ export default {
     uid() {
       return `${this.type}-${this.id}`;
     },
-    isThumbnail() {
-      return Array.isArray(this.asset.roles) && this.asset.roles.includes('thumbnail');
-    },
-    canShow() {
-      // We need to know the type, otherwise we don't even try to show it
-      if (typeof this.asset.type !== 'string') {
-        return false;
-      }
-      // If the tile renderer is a tile server, we can't really know what it supports so we pass all images
-      else if (this.tileRendererType === 'server' && imageMediaTypes.includes(this.asset.type)) {
-        return true;
-      }
-      // Don't pass GDAL VFS URIs to client-side tile renderer: https://github.com/radiantearth/stac-browser/issues/116
-      else if (this.isGdalVfs && this.tileRendererType === 'client') {
-        return false;
-      }
-      // Only http(s) links and relative links are supported
-      else if (!this.isBrowserProtocol) {
-        return false;
-      }
-      // Otherwise, all images that a browser can read are supported + JSON
-      else if (mapMediaTypes.includes(this.asset.type)) {
-        return true;
-      }
-      return false;
-    },
     fileFormat() {
       if (typeof this.asset.type === "string" && this.asset.type.length > 0) {
         return this.formatMediaType(this.asset.type);
@@ -180,72 +125,6 @@ export default {
         return this.formatMediaType(this.asset.type, null, {shorten: true});
       }
       return null;
-    },
-    protocol() {
-      if (typeof this.href === 'string') {
-        if (this.href) {
-          let match = this.href.match(/^(\w+):\/\//);
-          if (match) {
-            return match[1].toLowerCase();
-          }
-        }
-      }
-      return null;
-    },
-    isBrowserProtocol() {
-      return (!this.protocol && !this.isGdalVfs) || browserProtocols.includes(this.protocol);
-    },
-    isGdalVfs() {
-      return Utils.isGdalVfsUri(this.asset.href);
-    },
-    href() {
-      if (typeof this.asset.href !== 'string') {
-        return null;
-      }
-      let baseUrl = null;
-      if (this.context instanceof STAC) {
-        baseUrl = this.context.getAbsoluteUrl();
-      }
-      return this.getRequestUrl(this.asset.href, baseUrl);
-    },
-    from() {
-      if (this.isGdalVfs) {
-        let type = this.asset.href.match(/^\/vsi([a-z\d]+)(_streaming)?\//);
-        return this.protocolName(type);
-      }
-      else {
-        return this.protocolName(this.protocol);
-      }
-    },
-    browserCanOpenFile() {
-      if (this.isGdalVfs)  {
-        return false;
-      }
-      if (Utils.canBrowserDisplayImage(this.asset)) {
-        return true;
-      }
-      else if (typeof this.asset.type === 'string') {
-        switch(this.asset.type.toLowerCase()) {
-          case 'text/html':
-          case 'application/xhtml+xml':
-          case 'text/plain':
-          case 'application/pdf':
-            return true;
-        }
-      }
-      return false;
-    },
-    buttonText() {
-      if (this.browserCanOpenFile && this.isBrowserProtocol) {
-        return this.$t('open');
-      }
-      let where = (!this.isBrowserProtocol && this.from) ? 'withSource' : 'generic';
-      return this.$t(`assets.download.${where}`, {source: this.from});
-    },
-    copyButtonText() {
-      let what = this.isGdalVfs ? 'copyGdalVfsUrl' : 'copyUrl';
-      let where = (!this.isBrowserProtocol && this.from) ? 'withSource' : 'generic';
-      return this.$t(`assets.${what}.${where}`, {source: this.from});
     }
   },
   created() {
@@ -270,33 +149,7 @@ export default {
       return role;
     },
     show() {
-      let asset = Object.assign({}, this.asset);
-      // Override asset href with absolute URL if not a GDAL VFS
-      if (!this.isGdalVfs) {
-        asset.href = this.href;
-      }
-      this.$emit('show', asset, this.id, this.isThumbnail);
-    },
-    protocolName(protocol) {
-      if (typeof protocol !== 'string') {
-        return '';
-      }
-      switch(protocol.toLowerCase()) {
-        case 's3':
-          return this.$t('protocol.s3');
-        case 'abfs':
-        case 'abfss':
-          return this.$t('protocol.azure');
-        case 'gcs':
-          return this.$t('protocol.gcs');
-        case 'ftp':
-          return this.$t('protocol.ftp');
-        case 'oss':
-          return this.$t('protocol.oss');
-        case 'file':
-          return this.$t('protocol.file');
-      }
-      return '';
+      this.$emit('show', ...arguments);
     },
     collapseToggled(isVisible) {
       let event = isVisible ? 'openCollapsible' : 'closeCollapsible';
