+++ src/components/Link.vue
@@ -1,20 +1,16 @@
 <template>
   <li>
-    <StacLink :id="popoverId" :data="link" :fallbackTitle="fallbackTitle" />
-    <b-popover v-if="actions.length > 0" :target="popoverId" triggers="hover" placement="bottom">
-      <b-button-group vertical>
-        <b-button v-for="action of actions" v-bind="action.btnOptions" :key="action.id" variant="primary" @click="action.onClick">
-          <component v-if="action.icon" :is="action.icon" class="mr-1" />
-          {{ action.text }}
-        </b-button>
-      </b-button-group>
+    <StacLink :id="popoverId" :data="link" :fallbackTitle="fallbackTitle" class="pr-1" />
+    <b-popover :target="popoverId" triggers="hover" placement="right" container="#stac-browser">
+      <h6 class="small text-muted text-center">{{ $t('additionalActions') }}</h6>
+      <HrefActions vertical :data="link" size="sm" />
     </b-popover>
   </li>
 </template>
 
 <script>
+import HrefActions from './HrefActions.vue';
 import StacLink from './StacLink.vue';
-import LinkActions from '../../linkActions.config';
 import { BPopover } from 'bootstrap-vue';
 
 let linkId = 0;
@@ -23,6 +19,7 @@ export default {
   name: "Link",
   components: {
     BPopover,
+    HrefActions,
     StacLink
   },
   props: {
@@ -38,19 +35,10 @@ export default {
   computed: {
     popoverId() {
       return "popover-link-" + linkId;
-    },
-    actions() {
-      return Object.entries(LinkActions)
-        .map(([id, plugin]) => new plugin(this.link, id))
-        .filter(plugin => plugin.show);
-    },
+    }
   },
   beforeCreate() {
     linkId++;
   }
 };
 </script>
-
-<style>
-
-</style>
