+++ src/locales/en/texts.json
@@ -1,4 +1,8 @@
 {
+  "actions": {
+    "openIn": "Open in {service}"
+  },
+  "additionalActions": "Additional Actions",
   "additionalResources": "Additional Resources",
   "anonymized": {
     "title": "Anonymized"
@@ -157,9 +161,6 @@
     "last": "Last »",
     "next": "Next ›",
     "previous": "‹ Previous"
-  },
-  "actions": {
-    "openIn": "Open in {service}"
   },
   "poweredBy": "Powered by {link}",
   "protocol": {
